+++ src/CMakeLists.txt
@@ -3,5 +3,5 @@ cmake_minimum_required(VERSION 3.29)
 
 add_library(
         fibHeap
-        INTERFACE FibHeap.h
+        INTERFACE fibonacci_heap.h
 )
