+++ src/fibonacci_heap.h
@@ -6,7 +6,7 @@
 #include <expected>
 #include <vector>
 #include <memory>
-#include "Node.h"
+#include "fibonacci_heap_node.h"
 
 namespace kstmc {
     template <
